requirements.txt
fastapi>=0.115.2,<1.0
uvicorn[standard]>=0.24.0,<1.0
python-redmine==2.4.0
GitPython==3.1.40
pydantic>=2.7.2,<3.0
anyio>=4.5,<5.0
jinja2>=3.1.2,<4.0
